pyproject.toml
[tool.poetry]
name = "mysql-mcp-server"
version = "0.1.0"
description = "MySQL MCP Server implementation with FastMCP"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastmcp = "^2.0.0"
mysql-connector-python = "^8.0.33"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "server.main:main" 